rename vite to vite-dev
